refactor: update import paths for about components
