+++ src/components/about/about-text.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import React from 'react';
-import MarkdownRenderer from './markdown-renderer';
-import SubHeader from './sub-header';
+import MarkdownRenderer from '../markdown-renderer';
+import SubHeader from '../sub-header';
 
 import { abouts } from '@/config/about';
