+++ src/components/nav-bar/index.tsx
@@ -1,0 +1,3 @@
+import { NavBar } from "./nav-bar";
+
+export default NavBar;
