+++ src/components/nav-bar/nav-bar.tsx
@@ -28,5 +28,3 @@ export const NavBar: React.FC = () => {
     </nav>
   );
 }
-
-export default NavBar;
