+++ src/components/side-bar/index.tsx
@@ -1,0 +1,3 @@
+import { SideBar } from "./side-bar";
+
+export default SideBar;
